requirements.txt
azure-functions==1.4.0
azure-iot-hub==2.2.3
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
uamqp==1.2.12
urllib3==1.25.10
